pyproject.toml
[tool.poetry]
name = "devcon_hackathon_2022"
version = "0.1.0"
description = ""
authors = ["Redza Rosli <redza.rosli@cognite.com>"]

[tool.poetry.dependencies]
python = ">3.9.7,<4.0"
openai = ">=0.27,<0.28"
pdfplumber = "^0.7.6"
numpy = "^1.24.1"
python-dotenv = "^1.0.0"
msal = "^1.22.0"
cognite-sdk = "4.1.1"
streamlit = "^1.22.0"
pandas = "<2.0.1"
pytesseract = "^0.3.10"
pdf2image = "^1.16.3"
opencv-python = "^4.7.0.72"
matplotlib = "^3.7.1"
easyocr = "^1.6.2"

[tool.poetry.dev-dependencies]
ipykernel = "^6.22.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
